requirements.txt
# Stable Diffusion Image Editor
# Dependencies for AMD Radeon GPU with DirectML support

# Core ML Framework
torch>=2.0.0
torch-directml>=0.2.5.dev240914

# Stable Diffusion
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Hugging Face Models Download
hf_xet>=1.2.0

# Face Restoration
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0

# Web UI
gradio>=4.15.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Fine-tuning (Phase 3)
peft>=0.7.0
bitsandbytes>=0.41.0; platform_system != "Windows"

# Utilities
tqdm>=4.66.0
huggingface-hub>=0.20.0
psutil>=5.9.0  # Memory monitoring for OOM prevention

# Advanced Memory Optimizations
tomesd>=0.1.3  # Token Merging for Stable Diffusion
pytest>=7.4.0
